requirements.txt
# QiTlalli Healthcare Platform Dependencies

# Core Odoo Framework
odoo==17.0
psycopg2-binary==2.9.7
gunicorn==21.2.0

# Python Utilities
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.3

# Document Processing
Pillow==10.0.0
reportlab==4.0.4
pypdf2==3.0.1
xlrd==2.0.1
openpyxl==3.1.2

# Authentication & Security
python-ldap==3.4.3
passlib==1.7.4

# Internationalization
babel==2.12.1
lxml==4.9.3
python-dateutil==2.8.2
pytz==2023.3

# Google Cloud Platform (for deployment automation)
google-cloud-secret-manager>=2.16.0
google-cloud-storage>=2.10.0
google-cloud-sql>=1.12.0
google-cloud-run>=0.9.0
google-cloud-build>=3.15.0
google-auth>=2.23.0
google-cloud-resource-manager>=1.10.0